requirements.txt
ccxt>=1.60.0
pandas>=1.3.0
numpy>=1.20.0
pyyaml>=6.0
ta-lib>=0.4.24
matplotlib>=3.4.0
python-telegram-bot>=13.7
requests>=2.26.0
schedule>=1.1.0
colorama>=0.4.4
tqdm>=4.62.0
pytz>=2021.3